New "use" statement in template views
You can delcare some "assets_packages" in the "extra" block of your
"compser.json" to predegine a table of assets plugins that can be
"used" in the views - the "_use()" method will then load all required
assets automatically. See the examples of "jquery.tablesorter" and
"jquery.highlight" in the package manifest.
The layout and demo pages now use this feature.

Some new Smartphone icons are setted and included in the pages header
(if present in "www/icons/").

CSS styles are cleaned to be just what necessary.

Upgrade to version 0.0.5